% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\name{build.lvd}
\alias{build.lvd}
\title{build.lvd}
\usage{
build.lvd(load, displ, time = NULL)
}
\arguments{
\item{load}{A vector of numeric values specifying the load.}

\item{displ}{A vector of numeric values specifying the displacement.}

\item{time}{An optional vector of numeric values specifying the time.}
}
\value{
An object of class lvd (load versus displacement)
}
\description{
Builds an object of class .lvd from specified vectors
}
\examples{
l <- c(0,2,4,8,0)
d <- c(0,1,2,3,4)
t <- c(0,0.1,0.2,0.3,0.4)
myLvd <- build.lvd(load = l, displ = d, time = t)
}
