% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{energy.rate}
\alias{energy.rate}
\title{energy.rate}
\usage{
energy.rate(
  x,
  y,
  blade = "wedge",
  width,
  depth,
  crop = "auto",
  unit.area = "mm2",
  unit.load = "N"
)
}
\arguments{
\item{x}{An object of class 'lvd' for the first pass (= with specimen)}

\item{y}{An object of class 'lvd' for the second pass (= empty)}

\item{blade}{A character string indicating the type of blade used for the
test ('wedge' or 'scissors'). Default is 'wedge.'}

\item{width}{For the wedge test, a numeric indicating the width of the specimen}

\item{depth}{For the wedge test, a numeric indicating the depth of the fracture}

\item{crop}{Character string indicating whether the lvd data should be
cropped automatically ('auto') or manually ('manual'). Default is 'auto', which crops after the maximum was reached.}

\item{unit.area}{The unit in which the cross area was taken. Default is
square mm ("mm2")}

\item{unit.load}{The unit in which the load was measured. Default is N}
}
\value{
A numeric value for the specimen's energy release rate G
}
\description{
Compute the energy release rate in J/m2
}
