% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.plots.R
\name{lvd.crop}
\alias{lvd.crop}
\title{lvd.crop}
\usage{
lvd.crop(x, method = "after.max")
}
\arguments{
\item{x}{An object of class 'lvd'}

\item{method}{The method for cropping data, either 'before.max', 'after.max'
or 'manual'.}
}
\value{
A new 'lvd' object
}
\description{
Crop the range of lvd values to keep using an interactive plot
}
