% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\docType{class}
\name{lvd}
\alias{lvd}
\title{Load versus displacement object}
\description{
Creates an object of class lvd containing the load, the displacement and the
time for each observation.
}
\details{
This function creates an *lvd* object which can be used for
  calculating most parameters available in the flsr package.
}
\section{Slots}{

\describe{
\item{\code{load}}{A vector of numeric values specifying the load.}

\item{\code{displacement}}{A vector of numeric values specifying the displacement.}

\item{\code{time}}{An optional vector of numeric values specifying the time.}
}}

\seealso{
\code{\link{lvd}}

\code{\link{read.lvm}}
}
