% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\name{read.lvm}
\alias{read.lvm}
\title{read.lvm}
\usage{
read.lvm(file, center = FALSE, inv.displ = FALSE, inv.load = FALSE)
}
\arguments{
\item{file}{An .lvm file}

\item{center}{Logical indicating whether to move lvd data above 0}

\item{inv.displ}{Logical indicating whether the displacement axis should be
reversed (TRUE) or not (FALSE)}

\item{inv.load}{Logical indicating whether the load axis should be reversed
(TRUE) or not (FALSE)}
}
\value{
An object of class lvd (load versus displacement)
}
\description{
Loads a .lvm file (from Lucas' mechanical tester FLS2) into
  memory
}
\examples{
#myLvd <- read.lvm(file = <path>, center = TRUE)
}
