% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{toughness}
\alias{toughness}
\title{toughness}
\usage{
toughness(
  x,
  lng,
  area,
  crop = "auto",
  unit.area = "mm2",
  unit.load = "N",
  zero.out = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{x}{An object of class 'lvd'}

\item{lng}{A numeric indicating the initial specimen length}

\item{area}{A numeric indicating the initial specimen cross-section area}

\item{crop}{Character string indicating whether the lvd data should be
cropped automatically ('auto') or manually ('manual').}

\item{unit.area}{The unit in which the cross area was taken. Default is
square mm ("mm2")}

\item{unit.load}{The unit in which the load was measured. Default is N}

\item{zero.out}{Logical indicating whether the stress-strain curve should be
"zero-outed" before the conversion. NOTE : INTRODUIRE UN ALPHA POUR DECIDER
QUEL POURCENTAGE EXCLURE ENTRE 1% et 5%}
}
\value{
A numeric value for Young's elasticity modulus
}
\description{
Compute the energy per unit volume, or toughness, in J/m3
}
